refactor: split RFID utils into submodules
